trying to catch unintentional deletion of files

+++ jwst/outlier_detection/outlier_detection_step.py
@@ -185,7 +185,7 @@ class OutlierDetectionStep(Step):
                         median_path = blot_path.replace('blot', 'median')
 
                         for fle in [outlr_file, blot_path, median_path]:
-                            if os.path.isfile(fle):
+                            if os.path.isfile(fle) and 'tweakreg' not in fle:
                                 os.remove(fle)
                                 self.log.debug(f"    {fle}")
             else:
